refactor: ItemDrawer と FavoritesPanel を切り出し (#222)
- メニュー追加・編集ドロワーを _components/ItemDrawer.tsx に移動
- お気に入り空状態を FavoritesPanel に分離し MenuItemList から表示
- カート・保存用の PFC 換算を menu-item-pfc に集約

Made-with: Cursor
